Revert "Fix #4228"
This reverts commit 6940dc3adee8d0b05ec5b88d11ce839e2926d341.
