+++ org.mwc.debrief.lite/src/main/java/org/mwc/debrief/lite/map/GenericDragTool.java
@@ -13,7 +13,10 @@ import org.mwc.debrief.lite.gui.GeoToolMapProjection;
 
 import MWC.GUI.Layer;
 import MWC.GUI.Layers;
+import MWC.GUI.Shapes.FindNearest;
+import MWC.GUI.Shapes.DraggableItem.LocationConstruct;
 import MWC.GenericData.WorldLocation;
+import MWC.GenericData.WorldVector;
 
 public class GenericDragTool extends CursorTool
 {
@@ -67,7 +70,7 @@ public class GenericDragTool extends CursorTool
   /** how close we have to be (in screen pixels) to display
    * hotspot cursor 
    */
-  protected static double SCREEN_JITTER = 11;
+  private static double SCREEN_JITTER = 11;
   
   /**
    * the component we're going to drag
@@ -152,6 +155,76 @@ public class GenericDragTool extends CursorTool
 
     lastCursor = greenCursor;
     _mapPane.setCursor(greenCursor);
+  }
+  
+  @Override
+  public void onMouseMoved(MapMouseEvent ev)
+  {
+    super.onMouseMoved(ev);
+    
+    // try to determine if we're going over an item, to
+    // change the cursor
+    
+    // don't bother if we're already in a pan operation
+    if (!panning && !lastCursor.equals(draggingCursor))
+    {
+      panePos = mouseDelta(ev.getPoint());
+
+      final WorldLocation cursorLoc = _projection.toWorld(panePos);
+      // find the nearest editable item
+      final LocationConstruct currentNearest = new LocationConstruct();
+      final int num = layers.size();
+      for (int i = 0; i < num; i++)
+      {
+        final Layer thisL = layers.elementAt(i);
+        if (thisL.getVisible())
+        {
+          // find the nearest items, this method call will recursively pass down
+          // through
+          // the layers
+          // final Layer thisLayer,
+          FindNearest.findNearest(thisL, cursorLoc, panePos, currentNearest,
+              null, layers);
+        }
+      }
+
+      // Note - the following test does a distance check using world distance,
+      // which is quite unreliable,
+      
+      // did we find anything?
+      if (currentNearest.populated())
+      {
+        // generate a screen point from the cursor pos plus our distnace
+        // NOTE: we're not basing this on the target location - we may not have
+        // a
+        // target location as such for a strangely shaped object
+        final WorldLocation tgtPt =
+            cursorLoc.add(new WorldVector(Math.PI / 2,
+                currentNearest._distance, null));
+
+        // is it close enough
+        final Point tPoint = _projection.toScreen(tgtPt);
+
+        // get click point
+        Point cursorPos = ev.getPoint();
+        
+        // get distance of click point from nearest object, in screen coords
+        final double scrDist = tPoint.distance(cursorPos);
+
+        if (scrDist <= SCREEN_JITTER && !lastCursor.equals(greenCursor))
+        {
+          lastCursor = greenCursor;
+          _mapPane.setCursor(greenCursor);
+        }
+        else if ( scrDist > SCREEN_JITTER && !lastCursor.equals(normalCursor) )
+        {
+          lastCursor = normalCursor;
+          _mapPane.setCursor(normalCursor);
+        }
+      }
+    }
+    
+    
   }
 
   @Override
